requirements.txt
# MCP Server Requirements
agency-swarm>=2.0.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
notion-client>=2.0.0
python-dotenv>=1.0.0